% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasic_2016_rpkm}
\alias{tasic_2016_rpkm}
\title{RPKM matrix for single-cell transcriptomic datasets}
\format{A matrix with 1809 columns (cells) and 24,057 rows (genes)
\describe{
 \item{rows}{Each row corresponds to a single RefSeq gene symbol.}
 \item{columns}{Each column is named using a unique cell identifier that corresponds to the sample_id column in the tasic_2016_anno table.}
}}
\usage{
tasic_2016_rpkm
}
\description{
A matrix consisting of RPKM values for all 1809 cells
characterized in Tasic, et al. (2016).
}
\details{
Each row corresponds to a single gene symbol. Each column corresponds to a single cell.
}
\keyword{datasets}
